fix: close the six the review's third pass found
Two are the same arithmetic in two places, and two are the same window
on two readers.

The lazy postings decoder gains the two checks it was missing. A term's
block count is the one number in the postings file no unit checksum
reaches — it stands in front of the blocks rather than inside any of
them — so the terms index now hands each entry an extent as well as a
start, and decoding that stops short of the next term's entry is
refused. And a frequency is bounded by the document's own token count,
which docoff makes arithmetic; the eager decoder already enforced it
through a per-document budget that needs every term, and the
per-document ceiling does not.

Collection-wide token totals are summed in uint64. A segment's own
total is ranged against maxInt where it is decoded, which is all one
segment can be asked for; the sum across an incrementally committed
corpus is a different quantity, and on a 32-bit build a wrapped one is
a negative average length BM25 divides by. Merge's copy of that sum is
guarded too — there a wrap is not a bad score but a published segment
claiming a token count no walk of its documents reaches.

Commit takes the write lock before it reads the manifest. Merge holds
the same lock while it publishes gen+1 and prunes, so a snapshot taken
outside it could aim Commit's pre-write RemoveAll at the live merged
segment and then publish a manifest naming directories already gone.

Scrub re-reads a manifest a merge replaced, the way Open does, and both
now recognise a vanished *section file* as that same window: a prune
unlinks a segment's tree entry by entry, so a reader can find the
directory standing and a file inside it already gone.

And a no-op Commit still sweeps. Open documents unnamed debris as
costing disk until the next Commit, which was false exactly when it
mattered: nothing pending meant the orphan outlived every commit that
carried no document.

+++ pkg/engine/index.go
@@ -411,10 +411,18 @@ func (ix *Index) Stats() (docs int, avgDocLen float64) {
 // the whole corpus, and an average over only the documents added since the last
 // commit would rescale every score the moment a commit happened.
 func (ix *Index) avgDocLen() float64 {
-	n, total := len(ix.docs), ix.totalLen
+	// The sum is wider than its terms. Each segment's own total is ranged
+	// against maxInt where it is decoded, which is all a segment can be asked
+	// for — it knows nothing about the segments beside it. Incremental commit
+	// makes the collection-wide sum a different quantity: on a 32-bit build a
+	// handful of segments of a few hundred million tokens each overflow an int,
+	// and a wrapped total is a negative average length BM25 divides every score
+	// by. The document count needs no such care — readManifest already refuses a
+	// segment list running past maxDocCount.
+	n, total := len(ix.docs), uint64(ix.totalLen)
 	for _, s := range ix.segs {
 		n += s.count
-		total += s.totalLen
+		total += uint64(s.totalLen)
 	}
 	if n == 0 {
 		return 0
